+++ MoneyM/Shared/Models/FinancialSummary/FinancialSummaryModel.swift
@@ -8,10 +8,10 @@
 import Foundation
 
 struct FinancialSummaryModel {
-	var startingBalance: Int = 0
-    var expense: Int = 0
-    var income: Int = 0
-	var balance: Int {
+	var startingBalance: Double = 0
+    var expense: Double = 0
+    var income: Double = 0
+	var balance: Double {
 		startingBalance + income + expense
 	}
 }
